Removed another kernel dependency from compiler
Moved another use of traversal framework from Cypher compiler to remove kernel dependency
