+++ community/cypher/cypher-compiler-2.3/src/main/scala/org/neo4j/cypher/internal/compiler/v2_3/spi/QueryContext.scala
@@ -20,8 +20,9 @@
 package org.neo4j.cypher.internal.compiler.v2_3.spi
 
 import org.neo4j.cypher.internal.compiler.v2_3.InternalQueryStatistics
+import org.neo4j.cypher.internal.compiler.v2_3.pipes.matching.PatternNode
 import org.neo4j.cypher.internal.frontend.v2_3.SemanticDirection
-import org.neo4j.graphdb.{PropertyContainer, Relationship, Node}
+import org.neo4j.graphdb.{Path, PropertyContainer, Relationship, Node}
 import org.neo4j.kernel.api.constraints.{NodePropertyExistenceConstraint, RelationshipPropertyExistenceConstraint, UniquenessConstraint}
 import org.neo4j.kernel.api.index.IndexDescriptor
 
@@ -124,6 +125,14 @@ trait QueryContext extends TokenContext {
   def nodeGetDegree(node: Long, dir: SemanticDirection, relTypeId: Int): Int
 
   def nodeIsDense(node: Long): Boolean
+
+  // Legacy dependency between kernel and compiler
+  def variableLengthPathExpand(node: PatternNode,
+                               realNode: Node,
+                               minHops: Option[Int],
+                               maxHops: Option[Int],
+                               direction: SemanticDirection,
+                               relTypes: Seq[String]): Iterator[Path]
 }
 
 trait LockingQueryContext extends QueryContext {
